Restructure caching
