+++ aws/logs_monitoring/caching/cloudwatch_log_group_cache.py
@@ -1,12 +1,7 @@
+import os
 import boto3
-
-from caching.base_tags_cache import (
-    BaseTagsCache,
-    logger,
-    sanitize_aws_tag_string,
-    send_forwarder_internal_metrics,
-    should_fetch_log_group_tags,
-)
+from caching.base_tags_cache import BaseTagsCache
+from caching.common import sanitize_aws_tag_string, send_forwarder_internal_metrics
 from settings import (
     DD_S3_LOG_GROUP_CACHE_FILENAME,
     DD_S3_LOG_GROUP_CACHE_LOCK_FILENAME,
@@ -18,7 +13,7 @@ class CloudwatchLogGroupTagsCache(BaseTagsCache):
     CACHE_LOCK_FILENAME = DD_S3_LOG_GROUP_CACHE_LOCK_FILENAME
 
     def should_fetch_tags(self):
-        return should_fetch_log_group_tags()
+        return os.environ.get("DD_FETCH_LOG_GROUP_TAGS", "false").lower() == "true"
 
     def build_tags_cache(self):
         """Makes API calls to GetResources to get the live tags of the account's Lambda functions
@@ -30,14 +25,16 @@ class CloudwatchLogGroupTagsCache(BaseTagsCache):
         """
         new_tags = {}
         for log_group in self.tags_by_id.keys():
-            log_group_tags = get_log_group_tags(log_group)
+            log_group_tags = self._get_log_group_tags(log_group)
             # If we didn't get back log group tags we'll use the locally cached ones if they exist
             # This avoids losing tags on a failed api call
             if log_group_tags is None:
                 log_group_tags = self.tags_by_id.get(log_group, [])
             new_tags[log_group] = log_group_tags
 
-        logger.debug("All tags in Cloudwatch Log Groups refresh: {}".format(new_tags))
+        self.logger.debug(
+            "All tags in Cloudwatch Log Groups refresh: {}".format(new_tags)
+        )
         return True, new_tags
 
     def get(self, log_group):
@@ -54,45 +51,44 @@ class CloudwatchLogGroupTagsCache(BaseTagsCache):
         """
         if self._is_expired():
             send_forwarder_internal_metrics("local_cache_expired")
-            logger.debug("Local cache expired, fetching cache from S3")
+            self.logger.debug("Local cache expired, fetching cache from S3")
             self._refresh()
 
         log_group_tags = self.tags_by_id.get(log_group, None)
         if log_group_tags is None:
             # If the custom tag fetch env var is not set to true do not fetch
             if not self.should_fetch_tags():
-                logger.debug(
+                self.logger.debug(
                     "Not fetching custom tags because the env variable DD_FETCH_LOG_GROUP_TAGS is "
                     "not set to true"
                 )
                 return []
-            log_group_tags = get_log_group_tags(log_group) or []
+            log_group_tags = self._get_log_group_tags(log_group) or []
             self.tags_by_id[log_group] = log_group_tags
 
         return log_group_tags
 
-
-def get_log_group_tags(log_group):
-    response = None
-    try:
-        send_forwarder_internal_metrics("list_tags_log_group_api_call")
-        response = cloudwatch_logs_client.list_tags_log_group(logGroupName=log_group)
-    except Exception as e:
-        logger.exception(f"Failed to get log group tags due to {e}")
-    formatted_tags = None
-    if response is not None:
-        formatted_tags = [
-            (
-                "{key}:{value}".format(
-                    key=sanitize_aws_tag_string(k, remove_colons=True),
-                    value=sanitize_aws_tag_string(v, remove_leading_digits=False),
-                )
-                if v
-                else sanitize_aws_tag_string(k, remove_colons=True)
+    def _get_log_group_tags(self, log_group):
+        cloudwatch_logs_client = boto3.client("logs")
+        response = None
+        try:
+            send_forwarder_internal_metrics("list_tags_log_group_api_call")
+            response = cloudwatch_logs_client.list_tags_log_group(
+                logGroupName=log_group
             )
-            for k, v in response["tags"].items()
-        ]
-    return formatted_tags
-
-
-cloudwatch_logs_client = boto3.client("logs")
+        except Exception as e:
+            self.logger.exception(f"Failed to get log group tags due to {e}")
+        formatted_tags = None
+        if response is not None:
+            formatted_tags = [
+                (
+                    "{key}:{value}".format(
+                        key=sanitize_aws_tag_string(k, remove_colons=True),
+                        value=sanitize_aws_tag_string(v, remove_leading_digits=False),
+                    )
+                    if v
+                    else sanitize_aws_tag_string(k, remove_colons=True)
+                )
+                for k, v in response["tags"].items()
+            ]
+        return formatted_tags
