+++ aws/logs_monitoring/settings.py
@@ -264,3 +264,4 @@ DD_S3_STEP_FUNCTIONS_CACHE_LOCK_FILENAME = "step-functions-cache.lock"
 
 DD_TAGS_CACHE_TTL_SECONDS = int(get_env_var("DD_TAGS_CACHE_TTL_SECONDS", default=300))
 DD_S3_CACHE_LOCK_TTL_SECONDS = 60
+GET_RESOURCES_LAMBDA_FILTER = "lambda"
